requirements.txt
arcade @ https://github.com/pythonarcade/arcade/archive/refs/heads/development.zip
astar
pyyaml